requirements.txt
absl-py==0.15.0
affine==2.3.0
aiohttp==3.8.0
aiosignal==1.2.0
alabaster==0.7.12
alembic==1.4.1
allennlp==2.8.0
anyio==3.3.4
appdirs==1.4.4
argon2-cffi==21.1.0
arrow==1.2.1
astroid==2.6.6
async-timeout==4.0.0
atomicwrites==1.4.0
attrs==21.2.0
audioread==2.1.9
autopep8==1.6.0
Babel==2.9.1
backcall==0.2.0
backports.csv==1.0.7
base58==2.1.1
beautifulsoup4==4.10.0
binaryornot==0.4.4
black==21.10b0
bleach==4.1.0
blis==0.7.5
boto3==1.19.12
botocore==1.22.12
cached-path==0.3.2
cachetools==4.2.4
catalogue==2.0.6
certifi==2021.10.8
cffi==1.15.0
cftime==1.5.1.1
chardet==4.0.0
charset-normalizer==2.0.7
checklist==0.0.11
cheroot==8.5.2
CherryPy==18.6.1
click==8.0.3
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==2.0.0
configparser==5.1.0
cookiecutter==1.7.3
cryptography==35.0.0
cycler==0.11.0
cymem==2.0.6
Cython==0.29.24
dask==2021.10.0
dask-jobqueue==0.7.3
databricks-cli==0.16.2
datasets==1.15.1
debugpy==1.5.1
decorator==5.1.0
deepspeed==0.5.5
defusedxml==0.7.1
diff-match-patch==20200713
dill==0.3.4
distributed==2021.10.0
docker==5.0.3
docker-pycreds==0.4.0
docutils==0.17.1
entrypoints==0.3
et-xmlfile==1.1.0
fairscale==0.4.0
feedparser==6.0.8
filelock==3.3.2
Fiona==1.8.20
flake8==3.9.2
Flask==2.0.2
frozenlist==1.2.0
fsspec==2021.10.1
future==0.18.2
gensim==4.1.2
geopandas==0.10.2
gitdb==4.0.9
GitPython==3.1.24
google-api-core==2.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-cloud-core==2.1.0
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
graphviz==0.17
greenlet==1.1.2
grpcio==1.41.1
gspread==5.1.1
gunicorn==20.1.0
h5py==3.5.0
HeapDict==1.0.1
horovod==0.23.0
httplib2==0.20.2
huggingface-hub==0.1.1
idna==3.3
imageio==2.10.1
imagesize==1.2.0
imbalanced-learn==0.8.1
importlib-metadata==4.8.1
importlib-resources==5.4.0
inflection==0.5.1
iniconfig==1.1.1
intervaltree==3.1.0
ipykernel==6.5.0
ipython==7.29.0
ipython-genutils==0.2.0
iso-639==0.4.5
isort==5.10.0
itsdangerous==2.0.1
jaraco.classes==3.2.1
jaraco.collections==3.4.0
jaraco.functools==3.4.0
jaraco.text==3.6.0
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.0.2
jinja2-time==0.2.0
jmespath==0.10.0
joblib==1.1.0
json5==0.9.6
jsonnet==0.17.0
jsonpatch==1.32
jsonpointer==2.1
jsonschema==4.1.2
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyter-server==1.11.2
jupyterlab==3.2.1
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
keyring==23.2.1
kiwisolver==1.3.2
lazy-object-proxy==1.6.0
librosa==0.8.1
llvmlite==0.37.0
lmdb==1.2.1
locket==0.2.1
lxml==4.6.4
Mako==1.1.5
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
mlflow==1.21.0
more-itertools==8.10.0
mpi4py==3.1.1
msgpack==1.0.2
multidict==5.2.0
multiprocess==0.70.12.2
munch==2.5.0
murmurhash==1.0.6
mypy-extensions==0.4.3
mysql-connector-python==8.0.27
nbclassic==0.3.4
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
netCDF4==1.5.8
networkx==2.6.3
ninja==1.10.2.2
nltk==3.6.5
notebook==6.4.5
numba==0.54.1
numexpr==2.7.3
numpy==1.20.0
numpydoc==1.1.0
oauth2client==4.1.3
oauthlib==3.1.1
odfpy==1.4.1
opencv-python==4.5.4.58
openpyxl==3.0.9
overrides==3.1.0
packaging==21.2
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
partd==1.2.0
pathspec==0.9.0
pathtools==0.1.2
pathy==0.6.1
patternfork-nosql==3.6
pdfminer.six==20211012
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.0
pluggy==1.0.0
pooch==1.5.2
portend==3.0.0
poyo==0.5.0
preshed==3.0.6
prometheus-client==0.12.0
prometheus-flask-exporter==0.18.5
promise==2.3
prompt-toolkit==3.0.21
protobuf==3.19.1
psutil==5.8.0
ptyprocess==0.7.0
py==1.10.0
pyarrow==6.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.20
pydantic==1.9.0
pyDeprecate==0.3.1
pydocstyle==6.1.1
pydot==1.4.2
pyflakes==2.3.0
Pygments==2.10.0
pykeops==1.5
pylint==2.9.6
pyls-spyder==0.4.0
pynndescent==0.5.5
pyparsing==2.4.7
pyproj==3.2.1
pyprojroot==0.2.0
PyQt5==5.12.3
PyQt5-sip==12.9.0
PyQtWebEngine==5.12.1
pyrsistent==0.18.0
pytest==6.2.5
python-dateutil==2.8.2
python-docx==0.8.11
python-editor==1.0.4
python-lsp-black==1.0.0
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.2.4
python-slugify==5.0.2
pytorch-lightning==1.5.0
pytz==2021.3
PyWavelets==1.1.1
pyxdg==0.27
PyYAML==6.0
pyzmq==22.3.0
QDarkStyle==3.0.2
qstylizer==0.2.1
QtAwesome==1.1.0
qtconsole==5.1.1
QtPy==1.11.2
querystring-parser==1.2.4
rasterio==1.2.10
ray==1.8.0
redis==3.5.3
regex==2021.11.2
requests==2.26.0
requests-oauthlib==1.3.0
resampy==0.2.2
rope==0.21.0
rsa==4.7.2
Rtree==0.9.7
s3transfer==0.5.0
sacremoses==0.0.46
scikit-image==0.18.3
scikit-learn==1.0.1
scipy==1.7.1
seaborn==0.11.2
SecretStorage==3.3.1
Send2Trash==1.8.0
sentencepiece==0.1.96
sentry-sdk==1.4.3
sgmllib3k==1.0.0
Shapely==1.8.0
shortuuid==1.0.1
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
sniffio==1.2.0
snowballstemmer==2.1.0
snuggs==1.4.7
sortedcontainers==2.4.0
SoundFile==0.10.3.post1
soupsieve==2.3
spacy==3.1.4
spacy-legacy==3.0.8
Sphinx==4.2.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
spyder==5.1.5
spyder-kernels==2.1.3
SQLAlchemy==1.4.26
sqlitedict==1.7.0
sqlparse==0.4.2
srsly==2.4.2
subprocess32==3.5.4
tables==3.6.1
tabulate==0.8.9
tblib==1.7.0
tempora==4.1.2
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==1.8
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
text-unidecode==1.3
textdistance==4.2.2
thinc==8.0.13
threadpoolctl==3.0.0
three-merge==0.1.1
tifffile==2021.11.2
tinycss2==1.1.0
tokenizers==0.10.3
toml==0.10.2
tomli==1.2.2
toolz==0.11.1
torch==1.10.0+cu113
torchaudio==0.10.0+cu113
torchfile==0.1.0
torchmetrics==0.6.0
torchtext==0.11.0
torchvision==0.11.1+cu113
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
transformers==4.15.0
triton==1.1.1
typer==0.4.0
typing-extensions==3.10.0.2
ujson==4.2.0
umap-learn==0.5.2
urllib3==1.26.7
visdom==0.1.8.9
wandb==0.12.6
wasabi==0.8.2
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==2.0.2
wrapt==1.12.1
wurlitzer==3.0.2
xgboost==1.5.0
xlwt==1.3.0
xxhash==2.0.2
yapf==0.31.0
yarl==1.7.2
yaspin==2.1.0
zc.lockfile==2.0
zict==2.0.0
zipp==3.6.0
